+++ 3B2/3b2_sysdev.h
@@ -66,27 +66,27 @@ extern DEBTAB sys_deb_tab[];
 t_stat nvram_ex(t_value *vptr, t_addr exta, UNIT *uptr, int32 sw);
 t_stat nvram_dep(t_value val, t_addr exta, UNIT *uptr, int32 sw);
 t_stat nvram_reset(DEVICE *dptr);
-uint32 nvram_read(uint32 pa, uint8 size);
+uint32 nvram_read(uint32 pa, size_t size);
 t_stat nvram_attach(UNIT *uptr, char *cptr);
 t_stat nvram_detach(UNIT *uptr);
 const char *nvram_description(DEVICE *dptr);
-void nvram_write(uint32 pa, uint32 val, uint8 size);
+void nvram_write(uint32 pa, uint32 val, size_t size);
 
 /* 8253 Timer */
 t_stat timer_reset(DEVICE *dptr);
-uint32 timer_read(uint32 pa, uint8 size);
-void timer_write(uint32 pa, uint32 val, uint8 size);
+uint32 timer_read(uint32 pa, size_t size);
+void timer_write(uint32 pa, uint32 val, size_t size);
 t_stat timer_svc(UNIT *uptr);
 
 /* CSR */
 t_stat csr_ex(t_value *vptr, t_addr exta, UNIT *uptr, int32 sw);
 t_stat csr_dep(t_value val, t_addr exta, UNIT *uptr, int32 sw);
 t_stat csr_reset(DEVICE *dptr);
-uint32 csr_read(uint32 pa, uint8 size);
-void csr_write(uint32 pa, uint32 val, uint8 size);
+uint32 csr_read(uint32 pa, size_t size);
+void csr_write(uint32 pa, uint32 val, size_t size);
 
 /* TOD */
-uint32 tod_read(uint32 pa, uint8 size);
-void tod_write(uint32 pa, uint32 val, uint8 size);
+uint32 tod_read(uint32 pa, size_t size);
+void tod_write(uint32 pa, uint32 val, size_t size);
 
 #endif
